Random storage and memory tests
