+++ packages/solidity-debugger-core/test/memory.ts
@@ -1,73 +1,13 @@
 
-import {getShowValues, validMembers, createUserTypes, makeSimpleVariable, toStringObj, makeArrays, shuffle, generateTypes, range, generateRandomValues} from './storage';
-import {TypeName, Variable, getBytes} from '../src/artifacts/variables';
-import {printUserDefinedItems} from './storage';
-import {removeHexPrefix} from '../src/utils';
+import {toStringObj} from './helpers/state';
 import {decodeAssignment, State} from '../src/state';
 import {parseContractData} from '../src/artifacts/contracts';
-import parseTrace, {StepType} from '../src/trace';
+import parseTrace from '../src/trace';
 
 import newContract from './helpers/contract';
 import {testAdapter} from './helpers/adapter';
 
 import test from 'ava';
-
-const createMemoryVariable = (name: string, type: TypeName): Variable => ({
-    id: -1,
-    name: name,
-    location: 'memory', // actually not necessary
-    type: type,
-    scope: -1,
-    state: true,
-    bytes: getBytes(type),
-});
-
-// write some dont starts the variables, just write the values
-// so that we can make recursive calls to writesome
-// i.e. one struct creates an array
-function writeSome(name: string, type: TypeName, value: any) {
-    switch (type.name) {
-        case 'array':
-            return value.map((v, indx) => `${name}[${indx}] = ${v};`).join('\n')
-        case 'struct':
-            return (validMembers(type.members as Variable[])).map(v => writeSome(`${name}.${v.name}`, v.type, value[v.name])).join('\n')
-        case 'enum':
-            return `${name} = ${type.refName}.${value};`
-        case 'string':
-            return `${name} = "${value}";`
-        case 'bytes':
-            return `${name} = hex"${removeHexPrefix(value)}";`
-        default:
-            return `${name} = ${value};`
-    }
-}
-
-// maybe just do that part for all of them.
-function writeVariable(variable: Variable, value: any) {
-    const {type, name} = variable;
-
-    switch (type.name) {
-        case 'string':
-        case 'bytes':
-            return `
-                ${type.name} memory ${name};
-                ${writeSome(name, type, value)}
-            `
-        case 'array':
-            const base = type.base as TypeName;
-            return `
-                ${base.name}[] memory ${name} = new ${base.name}[](${value.length});
-                ${writeSome(name, type, value)}
-            `
-        case 'struct':
-            return `
-                ${variable.type.refName as string} memory ${name};
-                ${writeSome(name, type, toStringObj(getShowValues(value)))}
-            `
-        default:
-            throw Error(`Value not found ${type.name}`)
-    }
-}
 
 type Case = {
     source: string
@@ -202,7 +142,7 @@ const cases: Case[] = [
                 }
             }
         }
-    },
+    }
 ];
 
 async function applyCase(c: Case, t: any) {
@@ -221,108 +161,12 @@ async function applyCase(c: Case, t: any) {
 }
 
 for (const indx in cases) {
-    test(`cases0_${indx}`, async t => {
+    test(`memory_case_${indx}`, async t => {
         await applyCase(cases[indx], t);
     })
 }
 
-for (let i=0; i<1; i++) {
-    test(`random_${i}`, async t => {
-        await generate(t);
-    });
-}
-
-/*
-// Random
-(async() => {
-    // return;
-
-    await generate();
-})();
-*/
-
-async function generate(t) {
-
-    console.log("-- generate --")
-
-    const userTypes = createUserTypes(5);
-
-    console.log(userTypes)
-
-    // Create variable types
-
-    const avoid = [
-        'int',
-        'intx',
-        'uint',
-        'uintx',
-        'address',
-        'byte',
-        'bytesx'
-    ]
-    
-    let types: TypeName[] = shuffle([
-        ...range(0, 2).map(i => makeSimpleVariable(userTypes, avoid)),
-        ...generateTypes(2, makeArrays, userTypes),
-    ])
-
-    // Remove enum variables
-    types = types.filter(type => type.name != 'enum');
-
-    console.log(types)
-
-    // create variables to have a name reference
-
-    const variables = types.map((type, i) => createMemoryVariable(`val${i++}`, type));
-
-    console.log(JSON.stringify(variables, null, 4))
-
-    // create the results as the variable with a random value
-
-    let values: {[name: string]: any} = {}
-    let funcStr: string[] = [];
-    for (const variable of variables) {
-        const value = await generateRandomValues(variable.type);
-
-        values[variable.name] = value;
-        funcStr.push(writeVariable(variable, value))
-    }
-
-    let userTypesStr = printUserDefinedItems(userTypes)
-
-    const sample = `pragma solidity ^0.4.22;
-
-contract Sample {
-    ${userTypesStr}
-
-    function apply() {
-        ${funcStr.join('\n')}
-
-        int dummy = 0;
-    }
-}
-`
-
-    console.log(sample)
-
-    const retrieved = await run(sample)
-    
-    // values generated by random program
-    const cmp1 = toStringObj(getShowValues(values));
-
-    // values retrieved
-    const cmp2 = toStringObj(retrieved);
-
-    console.log("-- values --")
-    console.log(cmp1)
-
-    console.log("-- retrieved --")
-    console.log(cmp2)
-
-    t.deepEqual(cmp1, cmp2)
-}
-
-async function run(sample: string, method: string='apply', params: any[] = []) {
+export async function run(sample: string, method: string='apply', params: any[] = []) {
 
     const adapter = testAdapter(sample);
 
@@ -336,7 +180,7 @@ async function run(sample: string, method: string='apply', params: any[] = []) {
 
     let steps = await parseTrace(contracts, sources, transaction);
 
-    // One before the return statemenet
+    // One before the return statement
     const stmt = steps[steps.length - 2]
 
     const state = new State(transaction.blockNumber as number, true); // disable cache
